Cleanup Duplicate Code between classic and graphical ePOS-Print XML
