--- pretixprint/app/src/main/assets/demopage.eposprintxml
@@ -1,10 +1,0 @@
-<s:Envelope xmlns:s="http://schemas.xmlsoap.org/soap/envelope/">
-    <s:Body>
-        <epos-print xmlns="http://www.epson-pos.com/schemas/2011/03/epos-print">
-            <command>
-                546869732069732061207072657469785052494e542064656d6f207072696e74696e672e0d0a496620796f752063616e207265616420746869732c207072696e74696e6720776f726b73210d0a0d0a506c65617365207573652074686520666f6c6c6f77696e67206c696e65206f66206368617261637465727320746f2064657465726d696e65207468652070726f7065722070617065722077696474682e0d0a0d0a53656c65637420746865206c61737420636861726163746572206f6620746865206669727374207072696e746564206c696e652061732074686520225769647468206f6620726563656970742070617065722220696e20796f7572207072657469785052494e542073657474696e67732e0d0a0d0a4142434445464748494a4b4c4d4e4f505152535455565758595a303132333435363738396162636465666768696a6b6c6d6e6f707172737475767778797a0d0a0d0a0d0a0d0a0d0a1d56200d0a
-            </command>
-            <cut/>
-        </epos-print>
-    </s:Body>
-</s:Envelope>
