+++ pretixprint/app/src/main/java/eu/pretix/pretixprint/byteprotocols/GraphicePOSPrintXML.kt
@@ -27,7 +27,7 @@ class GraphicePOSPrintXML : CustomByteProtocol<Bitmap> {
         return type != "receipt"
     }
 
-    override fun convertPageToBytes(img: Bitmap,isLastPage: Boolean, previousPage: Bitmap?,conf: Map<String, String>, type: String): ByteArray {
+    override fun convertPageToBytes(img: Bitmap, isLastPage: Boolean, previousPage: Bitmap?,conf: Map<String, String>, type: String): ByteArray {
         return GraphicESCPOS().convertPageToBytes(img, isLastPage, previousPage, conf, type)
     }
 
@@ -39,69 +39,16 @@ class GraphicePOSPrintXML : CustomByteProtocol<Bitmap> {
         return Bitmap::class.java
     }
 
-    override fun sendNetwork(
-        host: String,
-        port: Int,
-        pages: List<CompletableFuture<ByteArray>>,
-        conf: Map<String, String>,
-        type: String,
-        context: Context
-    ) {
-        fun getSetting(key: String, def: String): String {
-            return conf[key] ?: context.defaultSharedPreferences.getString(key, def)!!
-        }
-
-        val deviceId = getSetting("hardware_${type}printer_deviceId", "local_printer")
-        val url = URL("http://$host:$port/cgi-bin/epos/service.cgi?devid=$deviceId&timeout=10000")
-
-        for (f in pages) {
-            with(url.openConnection() as HttpURLConnection) {
-                requestMethod = "POST"
-                setRequestProperty("Content-Type", "text/xml; charset=utf-8")
-                setRequestProperty("SOAPAction", "\"\"")
-
-                val wr = OutputStreamWriter(outputStream)
-                val escposdata = f.get(60, TimeUnit.SECONDS).toHex()
-                wr.write("""
-                    <s:Envelope xmlns:s="http://schemas.xmlsoap.org/soap/envelope/">
-                        <s:Body>
-                            <epos-print xmlns="http://www.epson-pos.com/schemas/2011/03/epos-print">
-                                <command>
-                                    $escposdata
-                                </command>
-                            </epos-print>
-                        </s:Body>
-                    </s:Envelope>
-                """.trimIndent())
-                wr.flush()
-                wr.close()
-
-                responseCode
-            }
-        }
+    override fun sendNetwork(host: String, port: Int, pages: List<CompletableFuture<ByteArray>>, conf: Map<String, String>, type: String, context: Context) {
+        ePOSPrintXML().sendNetwork(host, port, pages, conf, type, context)
     }
 
-    fun ByteArray.toHex(): String = joinToString(separator = "") { eachByte -> "%02x".format(eachByte) }
-
-    override fun sendUSB(
-        usbManager: UsbManager,
-        usbDevice: UsbDevice,
-        pages: List<CompletableFuture<ByteArray>>,
-        conf: Map<String, String>,
-        type: String,
-        context: Context
-    ) {
-        TODO("Not yet implemented")
+    override fun sendUSB(usbManager: UsbManager, usbDevice: UsbDevice, pages: List<CompletableFuture<ByteArray>>, conf: Map<String, String>, type: String, context: Context) {
+        ePOSPrintXML().sendUSB(usbManager, usbDevice, pages, conf, type, context)
     }
 
-    override fun sendBluetooth(
-        deviceAddress: String,
-        pages: List<CompletableFuture<ByteArray>>,
-        conf: Map<String, String>,
-        type: String,
-        context: Context
-    ) {
-        TODO("Not yet implemented")
+    override fun sendBluetooth(deviceAddress: String, pages: List<CompletableFuture<ByteArray>>, conf: Map<String, String>, type: String, context: Context) {
+        ePOSPrintXML().sendBluetooth(deviceAddress, pages, conf, type, context)
     }
 
     override fun allowedForConnection(type: ConnectionType): Boolean {
